requirements.txt
asgiref==3.10.0
azure-ai-inference==1.0.0b9
azure-core==1.36.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.36.0
dj-database-url==1.2.0
Django==5.2.7
django-cloudinary-storage==0.3.0
gunicorn==20.1.0
idna==3.11
isodate==0.7.2
psycopg2-binary==2.9.10
pypdf==6.2.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
whitenoise==6.5.0
